+++ operator/src/space/element.rs
@@ -1,5 +1,3 @@
-use crate::{FieldType, InnerProductSpace};
-
 use super::LinearSpace;
 
 /// Elements of linear spaces.
